Add utility functions for expected sums
There are multiple clue types that expect a sum on the cells it contains
which can share code: Roman Sums are just Killer Cages anyway, and all
Arrows can take a range of sums that needs to match. Region Sum Lines
can also use overlapping sums, but its current implementation is more
efficient. It might be worth it to move that to the `overlapping_sums`
function.

Signed-off-by: anderium <33520919+anderium@users.noreply.github.com>